chore(packages): move wallet menu to correct directory
