+++ packages/babylon-core-ui/src/widgets/sections/WalletMenu/components/WalletSettingItem.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Text } from '../../../components/Text';
+import { Text } from '../../../../components/Text';
 import { WalletToggle } from './WalletToggle';
 import { twJoin } from 'tailwind-merge';
